requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1
pyaudio==0.2.11
python-dotenv==1.0.0
websocket-client==1.6.4
